# harmonic-color-graph/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "EmbeddingRebuildPayload": {
        "additionalProperties": false,
        "properties": {
          "corpus_version": {
            "pattern": "^cv-[0-9]{4}-[0-9]{2}-[a-z0-9]{1,12}$",
            "title": "Corpus Version",
            "type": "string"
          }
        },
        "required": [
          "corpus_version"
        ],
        "title": "EmbeddingRebuildPayload",
        "type": "object"
      },
      "EmbeddingRebuildRequest": {
        "additionalProperties": false,
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/EmbeddingRebuildPayload"
          },
          "type": {
            "const": "embedding_rebuild",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "payload"
        ],
        "title": "EmbeddingRebuildRequest",
        "type": "object"
      },
      "EvaluationRunPayload": {
        "additionalProperties": false,
        "properties": {
          "song_limit": {
            "default": 100,
            "maximum": 20000.0,
            "minimum": 1.0,
            "title": "Song Limit",
            "type": "integer"
          },
          "suite": {
            "enum": [
              "keys",
              "labels"
            ],
            "title": "Suite",
            "type": "string"
          }
        },
        "required": [
          "suite"
        ],
        "title": "EvaluationRunPayload",
        "type": "object"
      },
      "EvaluationRunRequest": {
        "additionalProperties": false,
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/EvaluationRunPayload"
          },
          "type": {
            "const": "evaluation_run",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "payload"
        ],
        "title": "EvaluationRunRequest",
        "type": "object"
      },
...
      "GraphRebuildPayload": {
        "additionalProperties": false,
        "properties": {
          "corpus_version": {
            "pattern": "^cv-[0-9]{4}-[0-9]{2}-[a-z0-9]{1,12}$",
            "title": "Corpus Version",
            "type": "string"
          }
        },
        "required": [
          "corpus_version"
        ],
        "title": "GraphRebuildPayload",
        "type": "object"
      },
      "GraphRebuildRequest": {
        "additionalProperties": false,
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/GraphRebuildPayload"
          },
          "type": {
            "const": "graph_rebuild",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "payload"
        ],
        "title": "GraphRebuildRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Phase 1 harmonic data graph foundation API.",
    "title": "Harmonic Color Graph API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/jobs": {
      "post": {
        "operationId": "create_job_v2_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-HCG-Jobs-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Hcg-Jobs-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "embedding_rebuild": "#/components/schemas/EmbeddingRebuildRequest",
                    "evaluation_run": "#/components/schemas/EvaluationRunRequest",
                    "graph_rebuild": "#/components/schemas/GraphRebuildRequest"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/GraphRebuildRequest"
                  },
                  {
                    "$ref": "#/components/schemas/EmbeddingRebuildRequest"
                  },
                  {
                    "$ref": "#/components/schemas/EvaluationRunRequest"
                  }
                ],
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs-v2"
        ]
      }
    },
    "/v2/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v2_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-HCG-Jobs-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Hcg-Jobs-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs-v2"
        ]
      }
    },
    "/v2/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_v2_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-HCG-Jobs-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Hcg-Jobs-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs-v2"
        ]
      }
    },
    "/v2/jobs/{job_id}/events": {
      "get": {
        "operationId": "get_job_events_v2_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-HCG-Jobs-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Hcg-Jobs-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Events",
        "tags": [
          "jobs-v2"
        ]
      }
    },
    "/v2/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_v2_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-HCG-Jobs-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Hcg-Jobs-Token"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Job",
        "tags": [
          "jobs-v2"
        ]
      }
    },
...
  }
}